requirements.txt
streamlit
opencv-python
numpy
tensorflow
Pillow
matplotlib
ipywidgets
ipython
kagglehub
